1.信息部新增排名 2.新增运费维护模块

+++ server/src/main/java/com/dayuzl/coalapp/server/product/repository/ProductPriceRepository.java
@@ -8,6 +8,6 @@ import org.springframework.stereotype.Repository;
  * Created by 张杨 on 2017/5/23.
  */
 @Repository
-public interface ProductPriceRepository extends JpaRepository<ProductPrice, Long> {
+public interface ProductPriceRepository extends JpaRepository<ProductPrice, Integer> {
 
 }
